move test functionarity to AllQuestion
